major restructuring + cxsmiles coordinate generation
