+++ RanDepict/__init__.py
@@ -27,5 +27,7 @@ __all__ = [
     "RanDepict",
 ]
 
-
-from .randepict import RandomDepictor, RandomDepictorConfig, DepictionFeatureRanges, RandomMarkushStructureCreator
+from .config import RandomDepictorConfig
+from .depiction_feature_ranges import DepictionFeatureRanges
+from .randepict import RandomDepictor
+from .random_markush_structure_generator import RandomMarkushStructureCreator
